EWPP-2662: Moving test to Kernel.
