+++ modules/oe_link_lists_local/tests/src/FunctionalJavascript/LocalLinkListsTest.php
@@ -28,7 +28,6 @@ class LocalLinkListsTest extends WebDriverTestBase {
     'oe_link_lists_local_test',
     'entity_reference_revisions',
     'inline_entity_form',
-    'views',
   ];
 
   /**
