+++ modules/oe_link_lists_local/tests/src/Kernel/LocalLinkListsTest.php
@@ -6,7 +6,9 @@ namespace Drupal\Tests\oe_link_lists_local\Kernel;
 
 use Drupal\Core\Field\Entity\BaseFieldOverride;
 use Drupal\KernelTests\Core\Entity\EntityKernelTestBase;
+use Drupal\oe_link_lists\Entity\LinkList;
 use Drupal\oe_link_lists\Entity\LinkListInterface;
+use Drupal\views\Views;
 
 /**
  * Kernel tests for the local link lists.
@@ -24,6 +26,8 @@ class LocalLinkListsTest extends EntityKernelTestBase {
     'user',
     'system',
     'node',
+    'oe_link_lists_internal_source',
+    'views',
   ];
 
   /**
@@ -40,6 +44,7 @@ class LocalLinkListsTest extends EntityKernelTestBase {
     $this->installConfig([
       'oe_link_lists',
       'system',
+      'views',
     ]);
   }
 
@@ -134,6 +139,69 @@ class LocalLinkListsTest extends EntityKernelTestBase {
     $this->assertCount(1, $link_list_storage->loadMultiple([$link_list->id()]));
     $this->assertInstanceOf(LinkListInterface::class, $link_list_storage->load($link_list->id()));
     $this->assertInstanceOf(LinkListInterface::class, $link_list_storage->loadRevision($link_list->getRevisionId()));
+  }
+
+  /**
+   * Tests that views don't show the local link lists either.
+   */
+  public function testViewsQueryAlter(): void {
+    // Create 3 link lists: two normal ones (one with 0 value in the local
+    // field, and the other with NULL) and a local one.
+    $configuration = [
+      'source' => [
+        'plugin' => 'internal',
+        'plugin_configuration' => [
+          'entity_type' => 'node',
+          'bundle' => 'page',
+        ],
+      ],
+      'display' => [
+        'plugin' => 'title',
+      ],
+    ];
+
+    $link_list = LinkList::create([
+      'bundle' => 'dynamic',
+      'title' => 'The first visible link list',
+      'administrative_title' => 'The first visible link list',
+    ]);
+    $link_list->setConfiguration($configuration);
+    $link_list->save();
+
+    $link_list = LinkList::create([
+      'bundle' => 'dynamic',
+      'title' => 'The second visible link list',
+      'administrative_title' => 'The second visible link list',
+    ]);
+    $link_list->setConfiguration($configuration);
+    $link_list->set('local', NULL);
+    $link_list->save();
+
+    $link_list = LinkList::create([
+      'bundle' => 'dynamic',
+      'title' => 'The local link list',
+      'administrative_title' => 'The local link list',
+    ]);
+    $link_list->setConfiguration($configuration);
+    $link_list->set('local', TRUE);
+    $link_list->save();
+
+    $view = Views::getView('link_lists');
+    $view->setDisplay();
+    $view->preExecute();
+    $view->execute();
+    $results = $view->result;
+    $this->assertCount(2, $results);
+    $titles = [];
+    foreach ($results as $row) {
+      $entity = $row->_entity;
+      $titles[] = $entity->label();
+    }
+
+    $this->assertEquals([
+      'The first visible link list',
+      'The second visible link list',
+    ], $titles);
   }
 
 }
